[FIX] oca_pre_commit_hooks: Fix importing tests method from package
