+++ src/oca_pre_commit_hooks/checks_odoo_module.py
@@ -5,8 +5,6 @@ import os
 import sys
 from collections import defaultdict
 
-from common import get_checks_docstring
-
 from oca_pre_commit_hooks import checks_odoo_module_csv, checks_odoo_module_xml, utils
 
 DFTL_README_TMPL_URL = "https://github.com/OCA/maintainer-tools/blob/master/template/module/README.rst"  # noqa: B950
@@ -163,7 +161,7 @@ def lookup_manifest_paths(filenames_or_modules):
 
 def run(files_or_modules, enable=None, disable=None, no_verbose=False, no_exit=False, list_msgs=False):
     if list_msgs:
-        _, checks_docstring = get_checks_docstring(
+        _, checks_docstring = utils.get_checks_docstring(
             [ChecksOdooModule, checks_odoo_module_csv.ChecksOdooModuleCSV, checks_odoo_module_xml.ChecksOdooModuleXML]
         )
         print("Emittable messages with the current interpreter:", end="")
